修改赔率数据库的赔率类型

+++ src/main/java/digitalgame/service/OddsInfoService.java
@@ -3,7 +3,6 @@ package digitalgame.service;
 import digitalgame.model.po.OddsInfo;
 import digitalgame.model.po.UserBetInfo;
 
-import javax.xml.soap.SAAJResult;
 import java.util.HashMap;
 import java.util.List;
 
@@ -21,7 +20,7 @@ public interface OddsInfoService {
 
     int addOddsInfo(OddsInfo oddsInfo);
 
-    HashMap<String,Integer> selectOddsMap();
+    HashMap<String,Double> selectOddsMap();
 
     List<UserBetInfo> oddsNumber(int i, int j, int k, List<UserBetInfo> userBetInfoList, String resultDate);
 }
